+++ src/historia/project/__init__.py
@@ -1,4 +1,9 @@
-from .._add_to_project import add_to_project, transition_status, update_project_item_dates
+from .._add_to_project import (
+    add_to_project,
+    transition_status,
+    update_project_item_dates,
+    update_project_item_members,
+)
 from .._create_project import create_project_page
 
 __all__ = [
@@ -6,4 +11,5 @@ __all__ = [
     "create_project_page",
     "transition_status",
     "update_project_item_dates",
+    "update_project_item_members",
 ]